backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
openai>=1.30.0
anthropic>=0.28.0
python-dotenv>=1.0.1
httpx>=0.27.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
